requirements.txt
accelerate==1.1.0
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
decorator==4.4.2
deepgram-sdk==3.7.2
deprecation==2.1.0
distro==1.9.0
docx==0.2.4
exceptiongroup==1.2.2
fastapi==0.115.4
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
groq==0.22.0
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
lxml==5.3.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.12.1
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==2.1.2
openai==1.54.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.7
pooch==1.8.2
proglog==0.1.10
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.16
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.9.3
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.14.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
SpeechRecognition==3.14.2
starlette==0.41.2
streamlit==1.39.0
sympy==1.13.1
tenacity==9.0.0
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.20.3
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
tornado==6.4.1
tqdm==4.66.6
transformers==4.46.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
watchdog==5.0.3
websockets==13.1
yarl==1.17.1
google-generativeai
